add plotting scripts, touch up train file
