+++ notebooks_and_scripts/train.py
@@ -1,316 +1,382 @@
-"""RL training script for training RL agent, periodically evaluating it, and
-testing it on an entire period at the end. Uses 4 subprocesses to collect
-trajectories in parallel.
+"""RL training script.
+
+usage: train_v2.py [-h] -n EXPERIMENT_IDENTIFIER [-s SITE] [-r RANDOM_SEED]
+                   [-p PROJECT_ACTION]
+
+train models on EVChargingEnv
+
+optional arguments:
+  -h, --help            show this help message and exit
+  -n EXPERIMENT_IDENTIFIER, --experiment_identifier EXPERIMENT_IDENTIFIER
+  -s SITE, --site SITE  site of garage. caltech or jpl (default: caltech)
+  -r RANDOM_SEED, --random_seed RANDOM_SEED
+                        random seed (default: 0)
+  -p PROJECT_ACTION, --project_action PROJECT_ACTION
+                        whether to action-project in gym (default: True)
 """
+# Run 3x PPO discrete/continuous, A2C discrete/continuous, SAC
+from __future__ import annotations
 
 import argparse
-from argparse import RawTextHelpFormatter
 from datetime import datetime
 import gc
 import os
 import pickle
-from typing import Callable
+from typing import Callable, Optional, Union
 
+import gym
 import numpy as np
+import pandas as pd
 from stable_baselines3 import PPO, A2C, SAC
-from stable_baselines3.common.callbacks import EvalCallback, CallbackList
-from stable_baselines3.common.vec_env import SubprocVecEnv
+from stable_baselines3.common.base_class import BaseAlgorithm
+from stable_baselines3.common.callbacks import BaseCallback, EvalCallback
+from stable_baselines3.common.vec_env import SubprocVecEnv, VecEnv, sync_envs_normalization
 
-from sustaingym.envs.evcharging import EVChargingEnv, GMMsTraceGenerator, RealTraceGenerator
+from sustaingym.envs.evcharging import \
+    EVChargingEnv, GMMsTraceGenerator, RealTraceGenerator, DiscreteActionWrapper
 from sustaingym.algorithms.evcharging.baselines import RLAlgorithm
+from sustaingym.envs.evcharging.event_generation import AbstractTraceGenerator
+from sustaingym.envs.evcharging.utils import \
+    DATE_FORMAT, DEFAULT_PERIOD_TO_RANGE, DATE_FORMAT, SiteStr
+
 
 NUM_SUBPROCESSES = 4
-TIMESTEPS = 250_000
-EVAL_FREQ = 10_000 #10_000
-
-DATE_FORMAT = '%Y-%m-%d'
-FULL_PERIODS = {
-    'summer2019':   ('2019-05-01', '2019-08-31'),
-    'fall2019':     ('2019-09-01', '2019-12-31'),
-    'spring2020':   ('2020-02-01', '2020-05-31'),
-    'summer2021':   ('2021-05-01', '2021-08-31'),
-}
+TIMESTEPS = 20_000 # 250_000
+EVAL_FREQ = 10_000
 SAMPLE_EVAL_PERIODS = {
-    'summer2019':   ('2019-07-01', '2019-07-14'),
-    'fall2019':     ('2019-11-04', '2019-11-17'),
-    'spring2020':   ('2020-04-06', '2020-04-19'),
-    'summer2021':   ('2021-07-05', '2021-07-18'),
+    'Summer 2019':   ('2019-07-01', '2019-07-14'),
+    'Fall 2019':     ('2019-11-04', '2019-11-17'),
+    'Spring 2020':   ('2020-04-06', '2020-04-19'),
+    'Summer 2021':   ('2021-07-05', '2021-07-18'),
 }
 
 
-def num_days_in_period(full: bool, dp: str) -> int:
-    """Returns the number of days in period."""
-    d = FULL_PERIODS if full else SAMPLE_EVAL_PERIODS
-    dts = tuple(datetime.strptime(x, DATE_FORMAT) for x in d[dp])
-    td = dts[1] - dts[0]
-    return td.days + 1
+class EvalCallbackWithBreakdown(EvalCallback):
+    """Modifies EvalCallback's on_step to save reward breakdown during training."""
+    def __init__(
+        self,
+        eval_env: Union[gym.Env, VecEnv],
+        callback_on_new_best: Optional[BaseCallback] = None,
+        callback_after_eval: Optional[BaseCallback] = None,
+        n_eval_episodes: int = 5,
+        eval_freq: int = 10000,
+        log_path: Optional[str] = None,
+        best_model_save_path: Optional[str] = None,
+        verbose: int = 1,
+    ):
+        super().__init__(eval_env, callback_on_new_best, callback_after_eval,
+                         n_eval_episodes, eval_freq, log_path, best_model_save_path,
+                         True, False, verbose, True)
+        self.eval_env = eval_env
+        self.results = pd.DataFrame({})
+
+    def _on_step(self) -> bool:
+
+        continue_training = True
+
+        if self.eval_freq > 0 and self.n_calls % self.eval_freq == 0:
+
+            # Sync training and eval env if there is VecNormalize
+            if self.model.get_vec_normalize_env() is not None:
+                try:
+                    sync_envs_normalization(self.training_env, self.eval_env)
+                except AttributeError as e:
+                    raise AssertionError(
+                        "Training and eval env are not wrapped the same way, "
+                        "see https://stable-baselines3.readthedocs.io/en/master/guide/callbacks.html#evalcallback "
+                        "and warning above."
+                    ) from e
+
+            # Reset success rate buffer
+            self._is_success_buffer = []
+
+            algorithm = RLAlgorithm(self.eval_env, self.model)
+            results = algorithm.run(self.n_eval_episodes)
+            results['timestep'] = self.num_timesteps
+            self.results = pd.concat([self.results, results])
+
+            if self.log_path is not None:
+                # Save results to csv
+                self.results.to_csv(self.log_path + '.csv', compression='gzip', index=False)
+
+                # Save success log if present
+                if len(self._is_success_buffer) > 0:
+                    self.evaluations_successes.append(self._is_success_buffer)
+
+            rewards = np.array(results['reward'])
+            mean_reward, std_reward = np.mean(rewards), np.std(rewards)
+            self.last_mean_reward = mean_reward
+
+            if self.verbose > 0:
+                print(f"Eval num_timesteps={self.num_timesteps}, " f"episode_reward={mean_reward:.2f} +/- {std_reward:.2f}")
+            # Add to current Logger
+            self.logger.record("eval/mean_reward", float(mean_reward))
+
+            if len(self._is_success_buffer) > 0:
+                success_rate = np.mean(self._is_success_buffer)
+                if self.verbose > 0:
+                    print(f"Success rate: {100 * success_rate:.2f}%")
+                self.logger.record("eval/success_rate", success_rate)
+
+            # Dump log so the evaluation results are printed with the correct timestep
+            self.logger.record("time/total_timesteps", self.num_timesteps, exclude="tensorboard")
+            self.logger.dump(self.num_timesteps)
+
+            # Handle new best mean reward
+            if mean_reward > self.best_mean_reward:
+                if self.verbose > 0:
+                    print("New best mean reward!")
+                if self.best_model_save_path is not None:
+                    self.model.save(os.path.join(self.best_model_save_path, "best_model"))
+                self.best_mean_reward = mean_reward
+                # Trigger callback on new best model, if needed
+                if self.callback_on_new_best is not None:
+                    continue_training = self.callback_on_new_best.on_step()
+
+            # Trigger callback after every evaluation, if needed
+            if self.callback is not None:
+                continue_training = continue_training and self._on_event()
+
+        return continue_training
 
 
-def get_env(full: bool, real_trace: bool, dp: str, site: str, project_action_in_env: bool, seed: int=None) -> Callable:
+def parse_args() -> argparse.Namespace:
+    """Parse command line arguments."""
+    parser = argparse.ArgumentParser(
+        description='train models on EVChargingEnv',
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+    parser.add_argument('-n', '--experiment_identifier', required=True, type=int)  ##
+    parser.add_argument(
+        '-t', '--train', type=str, nargs='+', required=True,  ##
+        help="Season. 'Summer 2019', 'Fall 2019', 'Spring 2020', 'Summer 2021'")
+    parser.add_argument(
+        '-e', '--test', type=str, nargs='+', required=True,  ##
+        help="Season. 'Summer 2019', 'Fall 2019', 'Spring 2020', 'Summer 2021'")
+    parser.add_argument(
+        '-s', '--site', type=str, default='caltech',
+        help='site of garage. caltech or jpl')
+    parser.add_argument(
+        '-d', '--discrete', type=bool, default=False,
+        help='whether to use discretized actions, default continuous')
+    parser.add_argument(
+        '-m', '--model_name', type=str, required=True,  ##
+        help='type of model. SAC, PPO, or A2C. DQN or DDPG currently not supported')
+    parser.add_argument(
+        '-l', '--lr', type=float, default=3e-4,
+        help='learning rate')
+    parser.add_argument(
+        '-g', '--gamma', type=float, default=0.9999,
+        help='discount factor, between 0 and 1')
+    parser.add_argument(
+        '-o', '--log_dir', default='./logs',
+        help='directory for saving logs and models')
+    parser.add_argument(
+        '-r', '--random_seed', type=int, default=0,
+        help='random seed')
+    return parser.parse_args()
+
+
+def build_save_path(args: argparse.Namespace) -> str:
+    """Build path to saving."""
+    if not os.path.exists(args.log_dir):
+        print('Creating log directory at:', args.log_dir)
+        os.makedirs(args.log_dir)
+    
+    discrete_tag = '_discrete' if args.discrete else ''
+    save_path = f'{args.log_dir}/{args.model_name}{discrete_tag}/exp{args.experiment_identifier}'
+    if not os.path.exists(save_path):
+        os.makedirs(save_path)
+    else:
+        print(f'save path {save_path} already exists! Aborting')
+        exit()
+
+    # Save arguments
+    with open(os.path.join(save_path, 'args.pkl'), 'wb') as f:
+        pickle.dump(args, f)
+    return save_path
+
+
+def get_model_class(model_name: str, discrete: bool) -> BaseAlgorithm:
+    model_class: type
+    if model_name == 'A2C':
+        model_class = A2C
+    elif model_name == 'PPO':
+        model_class = PPO
+    elif model_name == 'SAC':
+        assert not discrete
+        model_class = SAC
+    else:
+        raise ValueError("Unsupported model.")
+    return model_class
+
+
+def setup_model(model_name: str, env: gym.Env, gamma: float, lr: float, discrete: bool) -> BaseAlgorithm:
+    model_class = get_model_class(model_name, discrete)
+    model = model_class(
+        policy='MultiInputPolicy',
+        env=env,
+        learning_rate=lr,
+        gamma=gamma,
+        verbose=1)
+    return model
+
+
+def get_env(full: bool, real_trace: bool, dp: str, site: SiteStr, discrete: bool = False, seed: int=None) -> Callable:
     """Return environment.
 
     Args:
         full: if True, use full season; otherwise, use sample 2 weeks
         real_trace: choice of generator
-        dp: 'summer2019', 'fall2019', 'spring2020', 'summer2021'
+        dp: 'Summer 2019', 'Fall 2019', 'Spring 2020', 'Summer 2021'
         site: 'caltech' or 'jpl'
-        project_action_in_env: whether action projection should be done
+        discrete: whether to wrap environment in discrete action wrapper
         seed: seed for GMMs generator
     
     Returns:
         Callable of environment
     """
-    d = FULL_PERIODS if full else SAMPLE_EVAL_PERIODS
-    date_period = d[dp]
+    date_period = DEFAULT_PERIOD_TO_RANGE[dp] if full else SAMPLE_EVAL_PERIODS[dp]
 
     def _get_env() -> EVChargingEnv:
         if real_trace:
-            gen = RealTraceGenerator(site, date_period, sequential=True)
+            gen: AbstractTraceGenerator = RealTraceGenerator(site, date_period)
         else:
             gen = GMMsTraceGenerator(site, date_period, seed=seed)
-        return EVChargingEnv(gen, action_type='discrete', project_action_in_env=project_action_in_env)
+        
+        if discrete:
+            return DiscreteActionWrapper(EVChargingEnv(gen))
+        else:
+            return EVChargingEnv(gen)
     return _get_env
 
 
-### Copied from stable-baselines3
-import os
-import warnings
-from abc import ABC, abstractmethod
-from typing import Any, Callable, Dict, List, Optional, Union
-
-import gym
-import numpy as np
-
-from stable_baselines3.common.callbacks import BaseCallback, EvalCallback
-from stable_baselines3.common import base_class  # pytype: disable=pyi-error
-from stable_baselines3.common.evaluation import evaluate_policy
-from stable_baselines3.common.vec_env import DummyVecEnv, VecEnv, sync_envs_normalization
-
-
-# class EvalCallbackWithBreakdown(EvalCallback):
-#     """
-#     Modifies EvalCallback's on_step to save reward breakdown during training.
-#     """
-
-#     def __init__(
-#         self,
-#         eval_env: Union[gym.Env, VecEnv],
-#         callback_on_new_best: Optional[BaseCallback] = None,
-#         callback_after_eval: Optional[BaseCallback] = None,
-#         n_eval_episodes: int = 5,
-#         eval_freq: int = 10000,
-#         log_path: Optional[str] = None,
-#         best_model_save_path: Optional[str] = None,
-#         verbose: int = 1,
-#         project_action: bool = True
-#     ):
-#         super().__init__(eval_env, callback_on_new_best, callback_after_eval,
-#                          n_eval_episodes, eval_freq, log_path, best_model_save_path,
-#                          True, False, verbose, True)
-#         self.eval_env = eval_env
-#         self.project_action = project_action
-
-#         self.breakdown_results = {
-#             'profit': [],
-#             'carbon_cost': [],
-#             'excess_charge': [],
-#         }
-
-#     def _on_step(self) -> bool:
-
-#         continue_training = True
-
-#         if self.eval_freq > 0 and self.n_calls % self.eval_freq == 0:
-
-#             # Sync training and eval env if there is VecNormalize
-#             if self.model.get_vec_normalize_env() is not None:
-#                 try:
-#                     sync_envs_normalization(self.training_env, self.eval_env)
-#                 except AttributeError as e:
-#                     raise AssertionError(
-#                         "Training and eval env are not wrapped the same way, "
-#                         "see https://stable-baselines3.readthedocs.io/en/master/guide/callbacks.html#evalcallback "
-#                         "and warning above."
-#                     ) from e
-
-#             # Reset success rate buffer
-#             self._is_success_buffer = []
-
-#             algorithm = RLAlgorithm(self.model, project_action=self.project_action)
-#             episode_rewards, breakdown = algorithm.run(self.n_eval_episodes, self.eval_env)
-
-#             if self.log_path is not None:
-#                 self.evaluations_timesteps.append(self.num_timesteps)
-#                 self.evaluations_results.append(episode_rewards)
-
-#                 for c in breakdown:
-#                     self.breakdown_results[c].append(breakdown[c])
-
-#                 kwargs = {}
-#                 # Save success log if present
-#                 if len(self._is_success_buffer) > 0:
-#                     self.evaluations_successes.append(self._is_success_buffer)
-#                     kwargs = dict(successes=self.evaluations_successes)
-
-#                 np.savez(
-#                     self.log_path,
-#                     timesteps=self.evaluations_timesteps,
-#                     results=self.evaluations_results,
-#                     profit=self.breakdown_results['profit'],
-#                     carbon_cost=self.breakdown_results['carbon_cost'],
-#                     excess_charge=self.breakdown_results['excess_charge'],
-#                     **kwargs,
-#                 )
-
-#             mean_reward, std_reward = np.mean(episode_rewards), np.std(episode_rewards)
-#             self.last_mean_reward = mean_reward
-
-#             if self.verbose > 0:
-#                 print(f"Eval num_timesteps={self.num_timesteps}, " f"episode_reward={mean_reward:.2f} +/- {std_reward:.2f}")
-#             # Add to current Logger
-#             self.logger.record("eval/mean_reward", float(mean_reward))
-
-#             if len(self._is_success_buffer) > 0:
-#                 success_rate = np.mean(self._is_success_buffer)
-#                 if self.verbose > 0:
-#                     print(f"Success rate: {100 * success_rate:.2f}%")
-#                 self.logger.record("eval/success_rate", success_rate)
-
-#             # Dump log so the evaluation results are printed with the correct timestep
-#             self.logger.record("time/total_timesteps", self.num_timesteps, exclude="tensorboard")
-#             self.logger.dump(self.num_timesteps)
-
-#             if mean_reward > self.best_mean_reward:
-#                 if self.verbose > 0:
-#                     print("New best mean reward!")
-#                 if self.best_model_save_path is not None:
-#                     self.model.save(os.path.join(self.best_model_save_path, "best_model"))
-#                 self.best_mean_reward = mean_reward
-#                 # Trigger callback on new best model, if needed
-#                 if self.callback_on_new_best is not None:
-#                     continue_training = self.callback_on_new_best.on_step()
-
-#             # Trigger callback after every evaluation, if needed
-#             if self.callback is not None:
-#                 continue_training = continue_training and self._on_event()
-
-#         return continue_training
-
-
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='Train Script', formatter_class=RawTextHelpFormatter)
-    parser.add_argument('--exp', type=int)
-    parser.add_argument('--seed', type=int)
-    parser.add_argument('--project_action_in_env', type=bool)
-    parser.add_argument('--model', help='Either `ppo` or `a2c')
-    parser.add_argument('--site', default='caltech', help='Either `caltech` or `jpl')
-    parser.add_argument('--train_period', help='Either `summer2019`, `fall2019`, `spring2020`, or `summer2021`')
-    parser.add_argument('--test_periods', nargs="+", help='Either `summer2019`, `fall2019`, `spring2020`, or `summer2021`')
-    args = parser.parse_args()
-
-    assert args.exp and args.seed and args.project_action_in_env and args.model
-    assert args.train_period and args.test_periods
-
-    # Set up save directory
-    exp_path = f'logs/{args.model}/exp_{args.exp}/'
-    save_path = os.path.join(os.getcwd(), exp_path)
-    os.makedirs(save_path, exist_ok=True)
-
-    # Save arguments
-    with open(os.path.join(save_path, 'args.pkl'), 'wb') as f:
-        pickle.dump(args, f)
-
-    # Print experiment information
-    start = datetime.now()
-    print("\n-----------------------")
-    print(f"Experiment: {args.exp}")
-    print(f"Save path: ", save_path)
-    print(f"Time: {start}")
-    with open(os.path.join(save_path, 'args.pkl'), 'rb') as f:
-        saved_args = vars(pickle.load(f))
-        print("--- Saved args ---")
-        for arg_name, arg_val in saved_args.items():
-            print(f'{arg_name}: {arg_val}')
-    print("-----------------------\n")
-
+def setup_envs(args: argparse.Namespace, save_path: str) -> \
+               tuple[gym.Env, EvalCallbackWithBreakdown]:
+    """Set up environment and evaluation callback."""
+    # Create train environments
+    assert len(args.train) == 2
+    train_period = ' '.join(args.train)
     train_envs = []
     for i in range(NUM_SUBPROCESSES):
-        # train on GMM over full period
-        # set seed slightly differently so the environments aren't exact
-        #  copies of each other
         train_envs.append(
-            get_env(True, False, args.train_period,
-                    args.site, args.project_action_in_env, seed=args.seed + i)
+            get_env(True, False, train_period, args.site, args.discrete, seed=args.random_seed + i)
         )
     train_envs = SubprocVecEnv(train_envs)
 
-    # use CallbackList in case other callbacks are to be added
-    callbacks = []
+    # Set up periodic evaluation
 
-    for test_period in args.test_periods:
-        # running evaluation environment, run on only 2-week subset to limit variability
-        # evaluate every 10000 timesteps
-        # test on eval date period and project action during evaluation
-        running_eval_env = get_env(False, True, test_period, args.site, True)()
+    # Run only on 2-week subset to limit variability
+    assert len(args.test) == 2
+    test_period = ' '.join(args.test)
+    running_eval_env = get_env(False, True, test_period, args.site, args.discrete)()
 
-        sp_running_eval = os.path.join(save_path, test_period)
-        os.makedirs(sp_running_eval, exist_ok=True)
+    # Create evaluation callback
+    eval_callback = EvalCallbackWithBreakdown(
+        running_eval_env, best_model_save_path=save_path,
+        n_eval_episodes=14,
+        eval_freq=EVAL_FREQ // NUM_SUBPROCESSES,
+        log_path=save_path,
+    )
 
-        eval_callback = EvalCallbackWithBreakdown(
-            running_eval_env, best_model_save_path=sp_running_eval,
-            log_path=sp_running_eval, eval_freq=EVAL_FREQ // NUM_SUBPROCESSES,
-            n_eval_episodes=14, project_action=False  # environment already projecting
-        )
+    return train_envs, eval_callback
 
-        callbacks.append(eval_callback)
-    callbacks = CallbackList(callbacks)
 
-    # Create model
-    models = {'ppo': PPO, 'a2c': A2C, 'sac': SAC }
-    model = models[args.model]('MultiInputPolicy', train_envs, seed=args.seed)
+def printout_before_train(args: argparse.Namespace,
+                          save_path: str,
+                          start: float,
+                          verbose: int = 1) -> None:
+    """Print experiment information befure training."""
+    if verbose > 0:
+        print("-" * 25)
+        print(f"Experiment: {args.experiment_identifier}")
+        print(f"Time: {start}")
+        with open(os.path.join(save_path, 'args.pkl'), 'rb') as f:
+            saved_args = vars(pickle.load(f))
+            print("Saved args: ")
+            for arg_name, arg_val in saved_args.items():
+                print(f'{arg_name}: {arg_val}')
+        print("-" * 25)
 
-    # Train model
-    print('Training model...\n')
-    model.learn(total_timesteps=TIMESTEPS, callback=callbacks)
-    end = datetime.now()
-    print("-----------------------")
+
+def printout_after_train(start: datetime, end: datetime) -> None:
+    """Print experiment information after training."""
+    print("-" * 25)
     print("\nTraining finished.")
     print(f"Time: {end}")
     print(f"Time elapsed for training: {end - start}\n")
-    print("-----------------------")
+    print("-" * 25)
 
-    # Load best model to run on test_period's    
+
+def num_days_in_period(full: bool, dp: str) -> int:
+    """Returns the number of days in period."""
+    d = DEFAULT_PERIOD_TO_RANGE if full else SAMPLE_EVAL_PERIODS
+    dts = tuple(datetime.strptime(x, DATE_FORMAT) for x in d[dp])
+    return (dts[1] - dts[0]).days + 1
+
+
+def test_rl(args: argparse.Namespace, save_path: str) -> None:
+    # Load best model
+    bm_path = os.path.join(save_path, 'best_model.zip')
+    best_model = get_model_class(args.model_name, args.discrete).load(bm_path)
+
+    # Create test environment
+    test_period = ' '.join(args.test)
+    test_env = get_env(True, True, test_period, args.site, args.discrete)()
+
+    # Evaluate
+    print(f'Evaluating {args.model_name} on {test_period}...\n')
+    rla = RLAlgorithm(test_env, best_model)
+    ndip = num_days_in_period(True, test_period)
+    test_results = rla.run(ndip)
+
+    # Save
+    test_results_save_path = os.path.join(save_path, 'test_results.csv')
+    test_results.to_csv(test_results_save_path, compression='gzip', index=False)
+
+
+def printout_after_test(train_start: datetime, train_end: datetime, test_end: datetime) -> None:
+    print("-" * 25)
+    print(f'Time: {test_end}')
+    print(f"Time elapsed for testing: {test_end - train_end}")
+    print(f"Total time elapsed: {test_end - train_start}")
+    print("-" * 25)
+
+
+def main() -> None:
+    """Train and evaluate sb3 RL agent.
+    
+    Creates new directory in '.logs/{model_name}/exp{experiment_identifier}'
+        args.pkl            pickled command-line arguments
+        best_model.zip      saved best model during training
+        evaluations.csv     2-week evaluations over training
+        test_results.csv    full-period test results after training
+    """
+    # Parse arguments
+    args = parse_args()
+    save_path = build_save_path(args)
+    print("-" * 25)
+    print(f'Saving model and logs to {save_path}')
+
+    # Create environment and callback
+    env, callback = setup_envs(args, save_path)
+
+    # Create model
+    model = setup_model(args.model_name, env, args.gamma, args.lr, args.discrete)
+
+    # Train model
+    train_start = datetime.now()
+    printout_before_train(args, save_path, train_start)
+    model.learn(total_timesteps=TIMESTEPS, callback=callback)
+    train_end = datetime.now()
+    printout_after_train(train_start, train_end)
     del model
     gc.collect()
-    bm_path = os.path.join(save_path, args.test_periods[0], 'best_model.zip')  # evaluate best model on first test period
-    best_model = models[args.model].load(bm_path)
-    # rl_raw = RLAlgorithm(best_model, project_action=False)  # TODO what to do here
-    rl_project = RLAlgorithm(best_model, project_action=True)  # TODO SAME
 
-    print(f'Testing model on {args.test_periods}...\n')
-    cnt = 0
-    for rl, lbl in zip([rl_project], ['projection']):
-        for test_period in args.test_periods:
-            cnt += 1
-            print(f'Evaluating {lbl} RL on {test_period} ({cnt}/{len(args.test_periods)})... ')
-            assert test_period in FULL_PERIODS
-            test_save_path = os.path.join(save_path, lbl, test_period)
-            os.makedirs(test_save_path, exist_ok=True)
-
-            test_env = get_env(True, True, test_period, args.site, False)()
-            num_eval = num_days_in_period(True, test_period)
-            rewards, breakdown = rl.run(num_eval, test_env)
-            results = {
-                'rewards': rewards,
-                'breakdown': breakdown
-            }
-
-            with open(os.path.join(test_save_path, 'test_results.pkl'), 'wb') as f:
-                pickle.dump(results, f)
+    # Test model
+    test_rl(args, save_path)
     test_end = datetime.now()
+    printout_after_test(train_start, train_end, test_end)
 
-    print("\n-----------------------")
-    print(f'Time: {test_end}')
-    print(f"Time elapsed for testing: {test_end - end}")
-    print(f"Total time elapsed: {test_end - start}")
-    print("-----------------------\n")
+
+if __name__ == '__main__':
+    main()
